+++ tests/unit/validateAttachmentSize.test.js
@@ -1,8 +1,10 @@
 require("../../lib/csn-runtime-extension")
 const cds = require("@sap/cds")
+const path = require("path")
+const app = path.resolve(__dirname, "../incidents-app")
+const { axios, POST, PUT } = cds.test(app)
 const { readFileSync } = cds.utils.fs
 const { join } = cds.utils.path
-const app = join(__dirname, "../incidents-app")
 const { validateAttachmentSize } = require("../../lib/generic-handlers")
 const { newIncident, withUser } = require("../utils/testUtils")
 const { POST, PUT } = withUser("alice", cds.test(app))